+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Other fixes and improvements:
 - Codex entries that are locked by relationship or storyline progress now have fitting hints associated with them.
 - Updated a number of boon icons with enhanced visuals.
 - Fixed: When dying in Tartarus after room exits have been unlocked, part of the room reward preview animation is overlayed over the blacked-out death screen.
+- Fixed: Some codex entries overflow the intended area in the Codex.
 - Fixed: Some dialogues are never eligible to play.
 - Fixed: The game may get into a bad state by not setting required flags correctly.
 <!--Releasenotes end-->